Prevent locale fallback in backend
If locale fallback is enabled, using a language other than the one used in frontend prevent to display it in backend. For example, if you have locale in german in frontend and want to use english in backend, the english won't be displayed as default language but used as fallback.